Add `replaceUrl` and `persist` plugins
