+++ code/ts/library/src/plugins/attributes/backend/fetchIndicator.ts
